requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
psycopg2-binary
celery[rabbitmq]
pydantic-settings
pgvector
ollama
httpx
beautifulsoup4
